Extensive reworking of correlated random number generation, changing the behavior. 1.  All distributions, including log-normal, now interpret input mean and sd as being     about the the actual random variable (formerly they were about the log of the rv     for the lognormal). 2.  Correlation is now achieved by matching percentiles.  The previous strategy     was ineffective for beta, and possibly for log normal.
Note that although the intention is that blocking by rows will induce correlations between
men and women if they are on separate lines, they are actually independent for now.  This
was the old behavior as well.

This also includes major reworking of the code to factor out different parts of the logic.
Untested.