Changes to file name specification in monte.conf.
This began with the observation that in my section header regex the (\S.*)\s*\]
would never match anything to the \s* since the .* would absorb it.  So the
captured name would include trailing spaces.  This is exactly the same problem
I ran into with (\.csv)? never capturing anything.

So I read up on look-ahead and look-behind assertions and experimented.
Leading to the current results.  User instructions updated appropriately.

During testing I had trouble getting the test program (not in this commit) running
under Node; VSCode kept treating it as Python.  launch.json got a new
configuration to fix this.

Various small tweaks to the docs.