Top-down changes in v4.
Make the master object an instance of a new MasterClass.
Return it to caller of this module, and provide hooks all the way down.
The object responds to iterate() and done().

Pass scenario argument all the way to database.

Update a comment to reflect greater Set ability of Node v22.
Fix: master.csvs -> master.sources.

Proper transaction handling still to do.