Upgrade better-sqlite3 and inquirer to the latest versions. UNTESTED!
Although inquirer v9 was only in ESM format, v10 and later were released in both
ESM and CJS.  And there seemed no good reason to keep it or better-sqlite pinned
to a major version, based on the release notes.

Update the README.md instructions, including adding better-sqlite3 to list of things to
install, and removing instructions to keep inquirer pinned.  Also updated general discussion
of modules in both Notes.md and v4.literate.

Move Javascript discussion that was not about module systems to separate subsections.