Big refactor of maestro.py into separate files. I believe I've retained the same functionality as before, but still no testing.
Create a Basics class to hold shared information so it is accessible across modules.
So the global variables have moved to this class.  Some interface changes as a result.

Ugly hack so I could load the code.

Various VSCode checkers don't grock my dynamic manipulation of sys.path; I altered a
vscode setting so it would know where to look.