Add TimerLog and subclasses to monitor how long the runs are taking.
Add maestro.md as  a guide for the advanced user of maestro.

Start ChangeLog entries for 3.5.0.

Modified runSims to send a brief message giving the start time
of an iteration.  This will allow capturing the gap between iterations
earlier.

Full list of types for the JSON messages in Notes.md; obsolete
information on types deleted from the mc help.