TerminalTimerLog now seems to work.
Changed base runSims.js so it outputs progress even on final iteration;
otherwise the logger doesn't know it's done.

Use create_task to start the monitor() process without waiting on it.

LogTimer
   * use get() to allow for possible missing keys.
   * fix logic of determining whether each run had finished at least one iteration.
   * correct for 1 based indices of runs in messages vs 0 based indices internally.
   * Fix fraction in reported days, and simplify duration format.
   * Improve logic determining if should print a report.

Note that an issue probably remains with the final report; it may wait
for the entire _delay.

Notes.md gains extensive To Do for parallel runs.