Make test more like my proposed code.  Kinda works.
The display is a bit messed up because the console.log(x) happens before I
respond to the prompt.  The callback in .then() is async.

Also, the description was never visible.

At the end, this is on the screen:

(pyenv) PS C:\Users\rdboylan\Documents\KBD\mccli-parallel\bin\js> node test
Hi

? select dat files to vary (Press <space> to select, <a> to toggle all, <i> to invert selection, and <enter> to proceed)
>( ) myfile: DANGEROUS{
  value: 'myfile',
  name: 'myfile: DANGEROUS',
  description: 'Changes to myfile ordinarily require manual calibration.'
? select dat files to vary myfile: DANGEROUS
got responses
[ 'myfile' ]