maestro: factor class definition Basics into its own file
The primary motivation was to avoid type checking complaints in prepare.py that
Basics was not defined (even though it was quoted!).

But moving it set off another round of complaints that some accessors on a
Basics instance were not defined, because they are set dynamically after
initialization in basics.py.  I defined them as None in the __init__,
and then got more type errors because None didn't support the operations I
was doing.  So I predefined them with insances of the appropriate types.

There was also an issue because some of the code used __file__ to navigate the
tree and basics.py is in a lower directory than maestro.py.

The type checker is happy with the code in this commit; I have yet to try running it.